requested changes: - add npmDevDeps - use npmDeps & npmDevDeps - rename 1-simple mod qux to foo - mv jest dependencies into JestModule - use CreateReactAppModule
