+++ example/javascriptlib/basic/1-simple/foo/src/foo.ts
